Use the subscribe/unsubscribe and callback stream api features as designed.
Subscribe returns a token 'client' which can be used to unsubscribe
and subscribe takes in the callback object (ctxt) which is the stream_handle.
This eliminates a lock and an idx lookup on every message delivered.
In closing a stream handle, unsubscribe client first or you can deadlock if
terminating during a message delivery.

In openstore, remove unnecessary lock on handle during create and
only subscribe the handle after it is completely ready.